requirements.txt
requests==2.32.3
python-dotenv==1.0.1
pandas==2.2.3
numpy==2.2.3
scikit-learn==1.6.1
